% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{addFreqBars}
\alias{addFreqBars}
\title{Title}
\usage{
addFreqBars(
  p,
  clusters,
  offset = 0.75,
  bar_length = 3,
  bar_width = 0.4,
  freq_labels = FALSE
)
}
\arguments{
\item{p}{a phylogenetic tree plot created from the ggtree() function}

\item{clusters}{a vector representing the cell type or cluster of each cell (can be character or numeric)}

\item{offset}{distance between the heatmap and frequency bars}

\item{bar_length}{length of bar with max frequency}

\item{bar_width}{width of each frequency bar}

\item{freq_labels}{boolean indicated whether or not to show frequency bar labels}
}
\description{
a function to add the frequency bars for each cluster
}
