% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{addTree}
\alias{addTree}
\title{Title}
\usage{
addTree(p, offset = 0.3, font_size = 2.5, hjust = 0)
}
\arguments{
\item{p}{a phylogenetic tree plot created from the ggtree() function}

\item{offset}{distance between leaf nodes on the tree and their labels}

\item{font_size}{font size of leaf labels}

\item{hjust}{horizontal justification as defined in ggplot2}
}
\description{
a function to create a skeleton tree diagram to display significance
testing results on each node
}
