% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{colourTree}
\alias{colourTree}
\title{Title}
\usage{
colourTree(
  tree,
  point_size = 1.5,
  high = "red",
  mid = "lightgrey",
  low = "blue"
)
}
\arguments{
\item{p}{a phylogenetic tree plot created from the ggtree() function
with p$data containing test statisic and p-}
}
\description{
Adding statistical test results onto the tree by using
colourful nodes and branches
Takes a ggtree object with test results for each node and returns
a ggtree graph object
}
