% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\docType{data}
\name{DeBiasi_COVID_CD8_samp}
\alias{DeBiasi_COVID_CD8_samp}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(COVIDSampleData)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(COVIDSampleData)
}
\references{
De Biasi et al. (2020) Nat Commun 11, 3434
(\href{https://www.nature.com/articles/s41467-020-17292-4}{Nature})
}
\keyword{datasets}
