% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{findChildren}
\alias{findChildren}
\title{Title}
\usage{
findChildren(tree)
}
\arguments{
\item{tree}{a ggtree object}
}
\value{
a ggtree object with the data containing a column with the clusters
contained in each node
}
\description{
Title
}
