% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{geometricMean}
\alias{geometricMean}
\title{geometricMean}
\usage{
geometricMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector containing numeric values}
}
\value{
geomtric mean of vector x
}
\description{
getCellGMeans helper function
}
