% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{getParentProp}
\alias{getParentProp}
\title{getParentProp}
\usage{
getParentProp(vars1, vars2, n_cells)
}
\arguments{
\item{vars1}{name of cell type, matching to column in n_cells}

\item{vars2}{name of parent cell type, matching to column in n_cells}

\item{n_cells}{matrix of counts of each cell type per sample}
}
\value{
a vector containing the proportions of cell type vars1 as a
percent of parent vars2 per sample
}
\description{
getCellProp helper function
}
