% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{getTotalProp}
\alias{getTotalProp}
\title{getTotalProp}
\usage{
getTotalProp(vars1, n_cells, n_cells_pat)
}
\arguments{
\item{vars1}{name of cell type, matching to column in n_cells}

\item{n_cells}{matrix of counts of each cell type per sample}

\item{n_cells_pat}{vector containing number of cells per sample}
}
\value{
a vector containing the proportions of cell type vars1 as a
percent of total per sample
}
\description{
getCellProp helper function
}
