% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{plotInteractiveHeatmap}
\alias{plotInteractiveHeatmap}
\title{Title}
\usage{
plotInteractiveHeatmap(
  testedTree,
  clust_med_df,
  clusters,
  svg_width = 13,
  svg_height = 9
)
}
\arguments{
\item{testedTree}{a ggtree object that has been run through the testTree
function}
}
\description{
This function takes a hierarchical tree which has been
tested for proportion to all and proportion to parent cluster
}
