% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{plotSigScatter}
\alias{plotSigScatter}
\title{Title}
\usage{
plotSigScatter(testedTree, scatter_tooltip, max_val)
}
\arguments{
\item{treeDat}{}
}
\value{
a ggplot object, containing test statistics from testing proportions relative to parent against
the test statistics from testing absolute proportions.
}
\description{
Title
}
