% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{runGLMMTests}
\alias{runGLMMTests}
\title{runGLMMTests}
\usage{
runGLMMTests(td, clusters, samples, classes, pos_class_name, neg_class_name)
}
\arguments{
\item{td}{a dataframe of data from ggtree object}

\item{clusters}{a vector representing the cell type or cluster of each cell
(can be character or numeric). If numeric, cluster names need to be consecutive
starting from 1.}

\item{samples}{a vector identifying the patient each cell belongs to}

\item{classes}{a vector containing the patient outcome/class each cell belongs to}
}
\value{
a dataframe with pvalues and test statistics
}
\description{
This function runs GLMM using the treekoR inputs across all nodes
of the hierarchical tree of clusters, adapted from the diffcyt workflow.
(Unable to get direction of test statistic currently)
}
