% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{testTree}
\alias{testTree}
\title{Title}
\usage{
testTree(
  phylo,
  exprs,
  clusters,
  samples,
  classes,
  pos_class_name = NULL,
  subjects = NULL,
  paired = FALSE
)
}
\arguments{
\item{paired}{}

\item{tree}{a ggtree object}

\item{cells}{a dataframe containing the clusters for each cell, the
sample id, the subject id (needed for paired tests), and the group
that the subject/sample belongs to}
}
\value{

}
\description{
This function takes a hierarchical tree of the cluster
medians of a cytometry dataset, and then uses this structure to perform
t-tests between conditions of patients testing for difference using the
proportion of cluster relative to sample's n and proportion of cluster
relative to sample's n of hierarchical parent cluster.
Takes a ggtree object and returns a ggtree object with testing results
appended in the data
}
