% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{addFreqBars}
\alias{addFreqBars}
\title{Title}
\usage{
addFreqBars(
  p,
  clusters,
  offset = 0.75,
  bar_length = 3,
  bar_width = 0.4,
  freq_labels = FALSE
)
}
\arguments{
\item{p}{a phylogenetic tree plot created from the ggtree() function}
}
\description{
a function to add the frequency bars for each cluster
}
