% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{addTree}
\alias{addTree}
\title{Title}
\usage{
addTree(
  p,
  offset = 0.3,
  width = 0,
  color = NULL,
  cluster_angle = 0,
  colnames = TRUE,
  colnames_position = "bottom",
  colnames_angle = NULL,
  colnames_level = NULL,
  colnames_offset_x = 0,
  colnames_offset_y = -0.55,
  font_size = 2.5,
  hjust = 0,
  legend_title = "Clusters"
)
}
\arguments{
\item{p}{a phylogenetic tree plot created from the ggtree() function}
}
\description{
a function to create a skeleton tree diagram to display significance
testing results on each node
}
