% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTree.R
\name{colourTree}
\alias{colourTree}
\title{Title}
\usage{
colourTree(
  tree,
  point_size = 1.5,
  high = "#00c434",
  low = "purple",
  mid = "ivory2"
)
}
\arguments{
\item{tree}{a tree plot created from the ggtree() function
with p$data containing test statisic and p-}

\item{point_size}{size of nodes in the tree}

\item{high}{colour for large values}

\item{low}{colour for low values}

\item{mid}{colour for middle values}
}
\description{
Adding statistical test results onto the tree by using
colourful nodes and branches
Takes a ggtree object with test results for each node and returns
a ggtree graph object
}
