% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\docType{data}
\name{DeBiasi_COVID_CD8_samp}
\alias{DeBiasi_COVID_CD8_samp}
\title{COVID-19 Sample data}
\format{
An object of class \code{"SingeCellExperiment"}
}
\source{
\href{https://flowrepository.org/id/FR-FCM-Z2N5}{FlowRepository}
}
\usage{
data(COVIDSampleData)
}
\description{
Data from a an experiment investigating T cell compositions between COVID-19
patients and healthy control. This data has been transformed using a arcsinh transform
using a co-factor of 5 and randomly subsetted
}
\examples{
data(COVIDSampleData)
}
\references{
De Biasi et al. (2020) Nat Commun 11, 3434
(\href{https://www.nature.com/articles/s41467-020-17292-4}{Nature})
}
\keyword{datasets}
