% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{getClusterTree}
\alias{getClusterTree}
\title{Title
This function takes a CATALYST sce with clusters and creates a hierarchical tree}
\usage{
getClusterTree(exprs, clusters, hierarchy_method = "average")
}
\value{

}
\description{
Title
This function takes a CATALYST sce with clusters and creates a hierarchical tree
}
