% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTree.R
\name{getTreeResults}
\alias{getTreeResults}
\title{Title}
\usage{
getTreeResults(testedTree, sort_by = "parent")
}
\arguments{
\item{testedTree}{a ggtree object outputed from testTree()}

\item{sort_by}{whether to sort by p-values testing via proportions to parent or
p-values testing via absolute proportions. Values can can be c(NA, "parent", "all")}
}
\value{

}
\description{
Title
}
